Add use case diagram (please read description)
I understand this is late, and you may dock marks. But I will just say that this is due to another group member who did not do their work on time, and the diagram had to be rushed. I understand there will be peer evaluation, but I just thought to note it here too.